Achievements Page should update now

+++ EFRAndroidFrontEndTest/EFRFrontEndTest2/Assets/UserObject.cs
@@ -22,7 +22,7 @@ namespace EFRFrontEndTest2.Assets
         public string Timestamp { get { return m_Timestamp; } set { m_Timestamp = value; } }
         public string Charity { get { return m_Charity; } set { m_Charity = value; } }
         public string FirstName { get { return m_FirstName; } set { m_FirstName = value; } }
-        public int MoneyEarned { get { return m_MoneyEarned; } set { m_MoneyEarned = value; } }
+        public double MoneyEarned { get { return m_MoneyEarned; } set { m_MoneyEarned = value; } }
         public int QuestionsAnswered { get { return m_QuestionsAnswered; } set { m_QuestionsAnswered = value; } }
         public string LastName { get { return m_LastName; } set { m_LastName = value; } }
         public string Username { get { return m_Username; } set { m_Username = value; } }       //These should only be changed when loading a user object (I used this implementation for readability and consistancy)
@@ -38,7 +38,7 @@ namespace EFRFrontEndTest2.Assets
         private string m_FirstName;
         private string m_LastName;
         private string m_Username;
-        private int m_MoneyEarned;
+        private double m_MoneyEarned;
         private int m_QuestionsAnswered;
         private int[] m_CompletedBlocks;
